Fix bug caused by caching in URLConnection.
There is a bug in the JavaFX Media library, caused by a problem in the
URLConnection class. The URLConnection class by default will use caches,
so if I edit a jar file, and put a new sound file inside, then the
URLConnection fails to see the new sound file, and the Media class plays
the old one. To fix this, I change the default use caches to always be
false using reflection in the main method of gui.Grow.

This problem was noticeable when the user dragged and dropped a sound
file into the program. The new sound file would be saved, but the
program would keep playing the old one. When the user reopened the
program, the program the played the new one.