Add ScoreDisplay action with prefix "s".
Also, change default behavior of action parsing. It used to be that an
action with no recognized prefix would be interpreted as a print.
However, consider what happens when the user types in "golf". They might
have intended this to be a print statement which prints "golf", but it
would instead be interpreted as a go action to the room "olf". By
requiring strict action syntax, the program eliminates this problem.